requirements.txt
# Python Dependencies für lokales Testen
# (AWS Lambda hat boto3 bereits installiert)

boto3>=1.26.0
requests>=2.28.0
python-dotenv>=1.0.0
